Made mathfield compatible with Py3 and Django>1.8
Fixed outdated imports

Added from_db_value in Mathfield

Updated katex to latest version from github

Added new line suport to admin preview

Changed renderMathFieldForm to set the html internally directly, this is
safer and easier than trying to escape the html in python and then
passing it to the function. The Signature of the function still remained
unchanged though.